Parser Improvements
Refactor Parsers
Improve UTM Parser
Add License to Project for Nuget
Update Nuget Pack Specs
