+++ CoordinateSharp/CoordinateSharp.csproj
@@ -47,7 +47,8 @@ For more information, please contact Signature Group, LLC at this address: sales
     <Description>A simple .NET standard library that is designed to assist with geographic coordinate conversions, formatting and location based solar and lunar calculations.</Description>
     <PackageReleaseNotes>Improves local time conversion effeciency and reduces benchmarks.</PackageReleaseNotes>
     <PackageTags>CoordinateSharp Latitude Longitude Coordinates Geography Sun Moon Solar Lunar Time MGRS UTM Julian ECEF Sunset Sunrise Set Moonset Moonrise</PackageTags>
-    <PackageLicenseExpression>AGPL-3.0-or-later</PackageLicenseExpression>
+    <!-- <PackageLicenseExpression>AGPL-3.0-or-later</PackageLicenseExpression>-->
+    <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
     <PackageIconUrl>https://raw.githubusercontent.com/Tronald/CoordinateSharp/master/ICON.png</PackageIconUrl>
     <PackageId>CoordinateSharp</PackageId>
     <Title>CoordinateSharp</Title>
@@ -81,7 +82,7 @@ For more information, please contact Signature Group, LLC at this address: sales
   <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netstandard2.0|AnyCPU'">
     <DocumentationFile>bin\Debug\netstandard2.0\CoordinateSharp.xml</DocumentationFile>
   </PropertyGroup>
-  
+
   <ItemGroup Condition="'$(TargetFramework)' == 'net40'">
     <Reference Include="System" />
     <Reference Include="System.Core" />
@@ -119,5 +120,8 @@ For more information, please contact Signature Group, LLC at this address: sales
     <PackageReference Include="System.Runtime.Serialization.Primitives">
       <Version>4.3.0</Version>
     </PackageReference>
+  </ItemGroup>
+  <ItemGroup>
+      <None Include="LICENSE.md" Pack="true" PackagePath="LICENSE.md"/>
   </ItemGroup>
 </Project>
