+++ CoordinateSharp/Parsers/CoordinatePart Formats/DegreeMinuteSecond.cs
@@ -1,0 +1,71 @@
+﻿/*
+CoordinateSharp is a .NET standard library that is intended to ease geographic coordinate 
+format conversions and location based celestial calculations.
+https://github.com/Tronald/CoordinateSharp
+
+Many celestial formulas in this library are based on Jean Meeus's 
+Astronomical Algorithms (2nd Edition). Comments that reference only a chapter
+are referring to this work.
+
+License
+
+Copyright (C) 2019, Signature Group, LLC
+  
+This program is free software; you can redistribute it and/or modify it under the terms of the GNU Affero General Public License version 3 
+as published by the Free Software Foundation with the addition of the following permission added to Section 15 as permitted in Section 7(a): 
+FOR ANY PART OF THE COVERED WORK IN WHICH THE COPYRIGHT IS OWNED BY Signature Group, LLC. Signature Group, LLC DISCLAIMS THE WARRANTY OF 
+NON INFRINGEMENT OF THIRD PARTY RIGHTS.
+
+This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY 
+or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details. You should have received a copy of the GNU 
+Affero General Public License along with this program; if not, see http://www.gnu.org/licenses or write to the 
+Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA, 02110-1301 USA, or download the license from the following URL:
+
+https://www.gnu.org/licenses/agpl-3.0.html
+
+The interactive user interfaces in modified source and object code versions of this program must display Appropriate Legal Notices, 
+as required under Section 5 of the GNU Affero General Public License.
+
+You can be released from the requirements of the license by purchasing a commercial license. Buying such a license is mandatory 
+as soon as you develop commercial activities involving the CoordinateSharp software without disclosing the source code of your own applications. 
+These activities include: offering paid services to customers as an ASP, on the fly location based calculations in a web application, 
+or shipping CoordinateSharp with a closed source product.
+
+For more information, please contact Signature Group, LLC at this address: sales@signatgroup.com
+*/
+using System;
+using System.Text.RegularExpressions;
+using System.Linq;
+using System.Globalization;
+
+namespace CoordinateSharp
+{
+    internal partial class FormatFinder_CoordPart
+    {
+        private static bool TryDegreeMinuteSecond(string s, out double[] d)
+        {
+            d = null;
+
+
+            double deg;
+            double min;
+            double sec;
+
+            string[] sA = SpecialSplit(s, true);
+            if (sA.Count() == 3)
+            {
+
+                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out deg))
+                { return false; }
+                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out min))
+                { return false; }
+                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out sec))
+                { return false; }
+
+                d = new double[] { deg, min, sec };
+                return true;
+            }
+            return false;
+        }
+    }
+}
