+++ CoordinateSharp/Parsers/FormatFinder.Coordinate.cs
@@ -40,7 +40,7 @@ using System.Globalization;
 
 namespace CoordinateSharp
 {
-    internal class FormatFinder
+    internal partial class FormatFinder
     {
         //Add main to Coordinate and tunnel to Format class. Add private methods to format.
         //WHEN PARSING NO EXCPETIONS FOR OUT OF RANGE ARGS WILL BE THROWN
@@ -210,418 +210,6 @@ namespace CoordinateSharp
             c = null;
             return false;
         }
-
-        #region PARSERS
-        private static bool TrySignedDegree(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 0) { return false; } //Should contain no letters
-
-            string[] sA = SpecialSplit(s, false);
-            double lat;
-            double lng;
-
-            double degLat;
-            double minLat; //Minutes & MinSeconds
-            double secLat;
-
-            int signLat = 1;
-
-            double degLng;
-            double minLng; //Minutes & MinSeconds
-            double secLng;
-
-            int signLng = 1;
-
-            switch (sA.Count())
-            {
-                case 2:
-                    if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out lat))
-                    { return false; }
-                    if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out lng))
-                    { return false; }
-                    d = new double[] { lat, lng };
-                    return true;
-                case 4:
-                    if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out degLat))
-                    { return false; }
-                    if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out minLat))
-                    { return false; }
-                    if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out degLng))
-                    { return false; }
-                    if (!double.TryParse(sA[3], NumberStyles.Any, CultureInfo.InvariantCulture, out minLng))
-                    { return false; }
-
-                    if (degLat < 0) { signLat = -1; }
-                    if (degLng < 0) { signLng = -1; }
-                    if (minLat >= 60 || minLat < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (minLng >= 60 || minLng < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    lat = (Math.Abs(degLat) + (minLat / 60.0)) * signLat;
-                    lng = (Math.Abs(degLng) + (minLng / 60.0)) * signLng;
-                    d = new double[] { lat, lng };
-                    return true;
-                case 6:
-                    if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out degLat))
-                    { return false; }
-                    if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out minLat))
-                    { return false; }
-                    if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out secLat))
-                    { return false; }
-                    if (!double.TryParse(sA[3], NumberStyles.Any, CultureInfo.InvariantCulture, out degLng))
-                    { return false; }
-                    if (!double.TryParse(sA[4], NumberStyles.Any, CultureInfo.InvariantCulture, out minLng))
-                    { return false; }
-                    if (!double.TryParse(sA[5], NumberStyles.Any, CultureInfo.InvariantCulture, out secLng))
-                    { return false; }
-                    if (degLat < 0) { signLat = -1; }
-                    if (degLng < 0) { signLng = -1; }
-                    if (minLat >= 60 || minLat < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (minLng >= 60 || minLng < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (secLat >= 60 || secLat < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (secLng >= 60 || secLng < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    lat = (Math.Abs(degLat) + (minLat / 60.0) + (secLat / 3600)) * signLat;
-                    lng = (Math.Abs(degLng) + (minLng / 60.0) + (secLng / 3600)) * signLng;
-                    d = new double[] { lat, lng };
-                    return true;
-                default:
-                    return false;
-            }
-        }
-        private static bool TryDecimalDegree(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 2) { return false; } //Should only contain 1 letter.
-
-            string[] sA = SpecialSplit(s, true);
-            if (sA.Count() == 2 || sA.Count() == 4)
-            {
-                double lat;
-                double lng;
-
-                double latR = 1; //Sets negative if South
-                double lngR = 1; //Sets negative if West
-
-                //Contact get brin directional indicator together with string
-                if (sA.Count() == 4)
-                {
-                    sA[0] += sA[1];
-                    sA[1] = sA[2] + sA[3];
-                }
-
-                //Find Directions
-                if (!sA[0].Contains("N") && !sA[0].Contains("n"))
-                {
-                    if (!sA[0].Contains("S") && !sA[0].Contains("s"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    latR = -1;
-                }
-                if (!sA[1].Contains("E") && !sA[1].Contains("e"))
-                {
-                    if (!sA[1].Contains("W") && !sA[1].Contains("w"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    lngR = -1;
-                }
-
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                sA[1] = Regex.Replace(sA[1], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out lat))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out lng))
-                { return false; }
-                lat *= latR;
-                lng *= lngR;
-                d = new double[] { lat, lng };
-                return true;
-            }
-
-            return false;
-        }
-        private static bool TryDegreeDecimalMinute(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 2) { return false; } //Should only contain 1 letter.
-
-            string[] sA = SpecialSplit(s, true);
-            if (sA.Count() == 4 || sA.Count() == 6)
-            {
-                double latD;
-                double latMS;
-                double lngD;
-                double lngMS;
-
-                double latR = 0; //Sets 1 if South
-                double lngR = 0; //Sets 1 if West
-
-                //Contact get in order to combine directional indicator together with string
-                //Should reduce 6 items to 4
-                if (sA.Count() == 6)
-                {
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; }
-                    else if (char.IsLetter(sA[2][0])) { sA[0] += sA[2]; }
-                    else { return false; }
-
-                    if (char.IsLetter(sA[3][0])) { sA[3] += sA[4]; sA[4] = sA[5]; }
-                    else if (char.IsLetter(sA[4][0])) { sA[3] += sA[4]; sA[4] = sA[5]; }
-                    else if (char.IsLetter(sA[5][0])) { sA[3] += sA[5]; }
-                    else { return false; }
-
-                    //Shift values for below logic
-                    sA[2] = sA[3];
-                    sA[3] = sA[4];
-                }
-
-                //Find Directions
-                if (!sA[0].Contains("N") && !sA[0].Contains("n") && !sA[1].Contains("N") && !sA[1].Contains("n"))
-                {
-                    if (!sA[0].Contains("S") && !sA[0].Contains("s") && !sA[1].Contains("S") && !sA[1].Contains("s"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    latR = 1;
-                }
-                if (!sA[2].Contains("E") && !sA[2].Contains("e") && !sA[3].Contains("E") && !sA[3].Contains("e"))
-                {
-                    if (!sA[2].Contains("W") && !sA[2].Contains("w") && !sA[3].Contains("W") && !sA[3].Contains("w"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    lngR = 1;
-                }
-
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                sA[1] = Regex.Replace(sA[1], "[^0-9.]", "");
-                sA[2] = Regex.Replace(sA[2], "[^0-9.]", "");
-                sA[3] = Regex.Replace(sA[3], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out latD))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out latMS))
-                { return false; }
-                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out lngD))
-                { return false; }
-                if (!double.TryParse(sA[3], NumberStyles.Any, CultureInfo.InvariantCulture, out lngMS))
-                { return false; }
-
-                d = new double[] { latD, latMS, latR, lngD, lngMS, lngR };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryDegreeMinuteSecond(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 2) { return false; } //Should only contain 1 letter.
-
-            string[] sA = SpecialSplit(s, true);
-            if (sA.Count() == 6 || sA.Count() == 8)
-            {
-                double latD;
-                double latM;
-                double latS;
-                double lngD;
-                double lngM;
-                double lngS;
-
-                double latR = 0; //Sets 1 if South
-                double lngR = 0; //Sets 1 if West
-
-                //Contact get in order to combine directional indicator together with string
-                //Should reduce 8 items to 6
-                if (sA.Count() == 8)
-                {
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[3][0])) { sA[0] += sA[3]; }
-                    else { return false; }
-
-                    if (char.IsLetter(sA[4][0])) { sA[4] += sA[5]; sA[5] = sA[6]; sA[6] = sA[7]; }
-                    else if (char.IsLetter(sA[5][0])) { sA[4] += sA[5]; sA[5] = sA[6]; sA[6] = sA[7]; }
-                    else if (char.IsLetter(sA[7][0])) { sA[4] += sA[7]; }
-                    else { return false; }
-
-                    //Shift values for below logic
-                    sA[3] = sA[4];
-                    sA[4] = sA[5];
-                    sA[5] = sA[6];
-                }
-
-                //Find Directions
-                if (!sA[0].Contains("N") && !sA[0].Contains("n") && !sA[2].Contains("N") && !sA[2].Contains("n"))
-                {
-                    if (!sA[0].Contains("S") && !sA[0].Contains("s") && !sA[2].Contains("S") && !sA[2].Contains("s"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    latR = 1;
-                }
-                if (!sA[3].Contains("E") && !sA[3].Contains("e") && !sA[5].Contains("E") && !sA[5].Contains("e"))
-                {
-                    if (!sA[3].Contains("W") && !sA[3].Contains("w") && !sA[5].Contains("W") && !sA[5].Contains("w"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    lngR = 1;
-                }
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                sA[1] = Regex.Replace(sA[1], "[^0-9.]", "");
-                sA[2] = Regex.Replace(sA[2], "[^0-9.]", "");
-                sA[3] = Regex.Replace(sA[3], "[^0-9.]", "");
-                sA[4] = Regex.Replace(sA[4], "[^0-9.]", "");
-                sA[5] = Regex.Replace(sA[5], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out latD))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out latM))
-                { return false; }
-                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out latS))
-                { return false; }
-                if (!double.TryParse(sA[3], NumberStyles.Any, CultureInfo.InvariantCulture, out lngD))
-                { return false; }
-                if (!double.TryParse(sA[4], NumberStyles.Any, CultureInfo.InvariantCulture, out lngM))
-                { return false; }
-                if (!double.TryParse(sA[5], NumberStyles.Any, CultureInfo.InvariantCulture, out lngS))
-                { return false; }
-
-                d = new double[] { latD, latM, latS, latR, lngD, lngM, lngS, lngR };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryMGRS(string s, out string[] mgrs)
-        {
-            mgrs = null;
-            //Attempt Regex Match
-            Regex regex = new Regex("[0-9]{1,2}[a-z,A-Z]{3}\\d+");
-            Match match = regex.Match(s);
-            if(match.Success)
-            {
-                //Extract Numbers for one string MGRS
-                regex = new Regex("\\d+");
-                MatchCollection matches = regex.Matches(s);
-
-                //IF character count of Easting Northing aren't even return false as precisions is unknown.
-                int splitSpot = matches[1].Value.Count();
-                if (splitSpot%2 == 0)
-                {
-                    
-                    string longZone = matches[0].Value;
-                    string eastingNorthing = matches[1].Value;
-
-                    //Extract Letters
-
-                    regex = new Regex("[a-z,A-Z]");
-                    matches = regex.Matches(s);
-                    string latZone = matches[0].Value;
-                    string identifier = matches[1].Value + matches[2].Value;
-
-                    //Split Easting and Northing Values
-                    string easting = eastingNorthing.Substring(0, (int)(eastingNorthing.Length / 2));
-                    string northing = eastingNorthing.Substring((int)(eastingNorthing.Length / 2), (int)(eastingNorthing.Length / 2));
-
-                    mgrs = new string[] { longZone, latZone, identifier, easting, northing };
-                    return true;
-                }
-
-            }
-            string[] sA = SpecialSplit(s, false);
-
-            if (sA.Count() == 4 || sA.Count() == 5)
-            {
-                double zone;
-                string zoneL;
-                string diagraph;
-                double easting;
-                double northing;
-
-                if (sA.Count() == 5)
-                {
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; sA[3] = sA[4]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; ; sA[3] = sA[4]; }
-                    else { return false; }
-                }
-                zoneL = new string(sA[0].Where(Char.IsLetter).ToArray());
-                if (zoneL == string.Empty) { return false; }
-                if (zoneL.Count() != 1) { return false; }
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                diagraph = sA[1];
-                if(diagraph.Count() != 2) { return false; }
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out zone))
-                { return false; }
-                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out easting))
-                { return false; }
-                if (!double.TryParse(sA[3], NumberStyles.Any, CultureInfo.InvariantCulture, out northing))
-                { return false; }
-
-                mgrs = new string[] { zone.ToString(), zoneL, diagraph, easting.ToString(), northing.ToString() };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryUTM(string s, out string[] utm)
-        {
-            utm = null;
-            string[] sA = SpecialSplit(s, false);
-            if (sA.Count() == 3 || sA.Count() == 4)
-            {
-                double zone;
-                string zoneL;
-                double easting;
-                double northing;
-
-                if (sA.Count() == 4)
-                {
-
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else { return false; }
-                }
-               
-                zoneL = new string(sA[0].Where(Char.IsLetter).ToArray());
-                if (zoneL == string.Empty) { return false; }
-                if(zoneL.Count() != 1) { return false; }
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out zone))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out easting))
-                { return false; }
-                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out northing))
-                { return false; }
-
-                utm = new string[] { zone.ToString(), zoneL, easting.ToString(), northing.ToString() };
-                return true;
-            }
-            return false;
-        }
-     
-        private static bool TryCartesian(string s, out double[] d)
-        {
-            d = null;
-            string[] sA = SpecialSplit(s, false);
-
-            if (sA.Count() == 3)
-            {
-                double x;
-                double y;
-                double z;
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out x))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out y))
-                { return false; }
-                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out z))
-                { return false; }
-                d = new double[] { x, y, z };
-                return true;
-            }
-            return false;
-        }
-        #endregion
 
         //KEEP DASHES FOR SIGNED AND CARTESIAN AS THEY ARE USED FOR NEGATVE VALUES
         private static string[] SpecialSplit(string s, bool removeDashes)
