+++ CoordinateSharp/Parsers/FormatFinder.CoordinatePart.cs
@@ -39,7 +39,7 @@ using System.Text.RegularExpressions;
 using System.Globalization;
 namespace CoordinateSharp
 {
-    internal class FormatFinder_CoordPart
+    internal partial class FormatFinder_CoordPart
     {
         //Add main to Coordinate and tunnel to Format class. Add private methods to format.
         //WHEN PARSING NO EXCPETIONS FOR OUT OF RANGE ARGS WILL BE THROWN
@@ -185,124 +185,6 @@ namespace CoordinateSharp
 
             return false;
         }
-
-        private static bool TrySignedDegree(string s, int t, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 0) { return false; } //Should contain no letters
-
-            string[] sA = SpecialSplit(s, false);
-            double deg;
-            double min; //Minutes & MinSeconds
-            double sec;
-
-            int sign = 1;
-            switch (sA.Count())
-            {
-                case 1:
-                    if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out deg))
-                    { return false; }
-                    d = new double[] { deg };
-                    return true;
-                case 2:
-                    if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out deg))
-                    { return false; }
-                    if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out min))
-                    { return false; }
-
-                    if (deg < 0) { sign = -1; }
-                    if (min >= 60 || min < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    d = new double[] { (Math.Abs(deg) + (min / 60.0)) * sign };
-                    return true;
-                case 3:
-                    if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out deg))
-                    { return false; }
-                    if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out min))
-                    { return false; }
-                    if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out sec))
-                    { return false; }
-                    if (min >= 60 || min < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (sec >= 60 || sec < 0) { return false; } //Handle in parser as degree will be incorrect.
-
-                    if (deg < 0) { sign = -1; }
-                    d = new double[] { (Math.Abs(deg) + (min / 60.0) + (sec / 3600.0)) * sign };
-                    return true;
-                default:
-                    return false;
-            }
-        }
-        private static bool TryDecimalDegree(string s, int direction, out double[] d)
-        {
-            d = null;
-            int sign = 1;
-            //S or W
-            if (direction == 2 || direction == 3)
-            {
-                sign = -1;
-            }
-            double coord;
-
-            string[] sA = SpecialSplit(s, true);
-
-            if (sA.Count() == 1)
-            {
-                if (!double.TryParse(s, NumberStyles.Any, CultureInfo.InvariantCulture, out coord))
-                { return false; }
-
-                coord *= sign;
-                d = new double[] { coord };
-                return true;
-            }
-
-            return false;
-        }
-        private static bool TryDegreeDecimalMinute(string s, out double[] d)
-        {
-            d = null;
-
-            double deg;
-            double minSec;
-
-
-            string[] sA = SpecialSplit(s,true);
-            if (sA.Count() == 2)
-            {
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out deg))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out minSec))
-                { return false; }
-
-                d = new double[] { deg, minSec };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryDegreeMinuteSecond(string s, out double[] d)
-        {
-            d = null;
-
-
-            double deg;
-            double min;
-            double sec;
-
-            string[] sA = SpecialSplit(s,true);
-            if (sA.Count() == 3)
-            {
-
-                if (!double.TryParse(sA[0], NumberStyles.Any, CultureInfo.InvariantCulture, out deg))
-                { return false; }
-                if (!double.TryParse(sA[1], NumberStyles.Any, CultureInfo.InvariantCulture, out min))
-                { return false; }
-                if (!double.TryParse(sA[2], NumberStyles.Any, CultureInfo.InvariantCulture, out sec))
-                { return false; }
-
-                d = new double[] { deg, min, sec };
-                return true;
-            }
-            return false;
-        }
-
         private static int Find_Position(string s)
         {
             //N=0
